rename dataset related files; move lc dataset to separate file
